pdcl-13739: rename the auto validate file to be in line with how tests scope the function name. Fix an eslint issue
